Partition into modules
