+++ src/models/mod.rs
@@ -1,0 +1,5 @@
+mod ticket_draft;
+mod ticket;
+
+pub use ticket::*;
+pub use ticket_draft::*;
